Image resizing and Layout Update
Several things! First images are now resized in images.inc.php, there
is a set dimensions that they must meet, but I'll probably make it
dynamic at some point. Also, the book generates a "black" image for for
the resized image to reside. This is problematic because PNG files will
show a black background != good. So I added a private function in the
object to create a black picture, then clear it to save the PNG in a
clear which does = good! Also cleaned up the CSS and made things a bit
more dynamic as well…